Use size_t rather than int for the data structures. Export vector, deque and ring_buffer for testing and use in programs. Remove some inlines for now since these result in lots of compile warnings when also exported, the compiler is probably smart enough anyway.

+++ lib/data-struct/sliding_window.h
@@ -1,4 +1,5 @@
 #include <stdint.h>
+#include <stddef.h>
 
 #ifndef LIBTRACE_SLIDING_WINDOW_H
 #define LIBTRACE_SLIDING_WINDOW_H
@@ -9,13 +10,13 @@
 // All of start, elements and end must be accessed in the listed order
 // if LIBTRACE_RINGBUFFER_SPINNING is to work.
 typedef struct libtrace_slidingwindow {
-	volatile int start;
-	int size;
+	volatile size_t start;
+	size_t size;
 	volatile uint64_t start_number; 
 	void *volatile*elements;
 } libtrace_slidingwindow_t;
 
-void libtrace_slidingwindow_init(libtrace_slidingwindow_t * sw, int size, uint64_t start_number);
+void libtrace_slidingwindow_init(libtrace_slidingwindow_t * sw, size_t size, uint64_t start_number);
 inline void libtrace_zero_slidingwindow(libtrace_slidingwindow_t * sw);
 void libtrace_slidingwindow_destroy(libtrace_slidingwindow_t * sw);
